moved single and multpleselection to their own files
